Updated to preview What I have on the left side
I updated the previously yellow column in HomeScreenForTabletV2 to showcase what is in the list, though currently it is just showing what is in the grocery list as we don't have the others. In modifying that, I also ended up modifying the ListCard and GroceryListPage a little as they are interconnected.